Project : Web scraping of Flipkart Mobil data under > 50000Rs
Web scraping is the process of automatically extracting data from websites. It involves fetching web pages, parsing the HTML or other structured data on the page, and then extracting the desired information. Web scraping can be a powerful tool for collecting data from the internet for various purposes, such as research, analysis, data aggregation, and more.